Convert game gateway into service
