+++ backend/src/db_interactions_modules/game/classes/Score.ts
@@ -1,0 +1,8 @@
+export class Score {
+  player1: number
+  player2: number
+  constructor() {
+    this.player1 = 0
+    this.player2 = 0
+  }
+}
